Move discovery accounts button to developer settings
