+++ packages/extension/src/ui/features/accounts/AccountScreenEmpty.tsx
@@ -1,8 +1,6 @@
 import { Empty, EmptyButton, icons } from "@argent/ui"
 import { partition } from "lodash-es"
-import { FC, useEffect, useState } from "react"
-import { discoverAccounts } from "../../services/backgroundAccounts"
-import { LoadingScreen } from "../actions/LoadingScreen"
+import { FC, useEffect } from "react"
 import { useCurrentNetwork } from "../networks/useNetworks"
 import { AccountNavigationBar } from "./AccountNavigationBar"
 import { isHiddenAccount, useAccounts } from "./accounts.state"
@@ -11,7 +9,7 @@ import { autoSelectAccountOnNetwork } from "./switchAccount"
 import { AccountContainer } from "./AccountContainer"
 import { AccountTokens } from "../accountTokens/AccountTokens"
 
-const { WalletIcon, AddIcon, SearchIcon } = icons
+const { WalletIcon, AddIcon } = icons
 
 export interface AccountScreenEmptyProps {
   onAddAccount: () => void
@@ -21,8 +19,6 @@ export const AccountScreenEmpty: FC<AccountScreenEmptyProps> = ({
   onAddAccount,
 }) => {
   const currentNetwork = useCurrentNetwork()
-  const [discoveringAccount, setDiscoveringAccount] = useState(false)
-  const [accountDiscoveryFinished, setAccountDiscoveryFinished] = useState(false)
   const allAccounts = useAccounts({ showHidden: true })
   const [hiddenAccounts, visibleAccounts] = partition(
     allAccounts,
@@ -31,29 +27,14 @@ export const AccountScreenEmpty: FC<AccountScreenEmptyProps> = ({
   const hasVisibleAccounts = visibleAccounts.length > 0
   const hasHiddenAccounts = hiddenAccounts.length > 0
 
-  const onDiscoverAccounts = () => {
-    setDiscoveringAccount(true)
-    discoverAccounts(currentNetwork.id)
-      .then(() => {
-        setDiscoveringAccount(false)
-        setAccountDiscoveryFinished(true)
-      })
-      .catch((e) => {
-        console.error(e)
-        setDiscoveringAccount(false)
-      })
-  }
-
   useEffect(() => {
     /** User made some account visible then returned to this screen */
     if (hasVisibleAccounts) {
       autoSelectAccountOnNetwork(currentNetwork.id)
     }
-  }, [currentNetwork.id, hasVisibleAccounts, accountDiscoveryFinished])
+  }, [currentNetwork.id, hasVisibleAccounts])
 
-  if (allAccounts.length === 0 && discoveringAccount) {
-    return <LoadingScreen />
-  } else if (allAccounts.length > 0) {
+  if (allAccounts.length > 0) {
     const account = allAccounts[0]
     return (
       <AccountContainer
@@ -74,35 +55,12 @@ export const AccountScreenEmpty: FC<AccountScreenEmptyProps> = ({
       >
         <EmptyButton
           size={"sm"}
-          colorScheme={"transparent"}
-          color={"neutrals.500"}
           leftIcon={<AddIcon />}
           onClick={onAddAccount}
           loadingText={"Creating"}
         >
           New account
         </EmptyButton>
-        {
-          accountDiscoveryFinished ? (
-            <>
-              No active accounts found
-            </>
-          ) :
-            (<>
-              or
-              <EmptyButton
-                size={"sm"}
-                colorScheme={"transparent"}
-                color={"neutrals.500"}
-                leftIcon={<SearchIcon />}
-                onClick={onDiscoverAccounts}
-                loadingText={"Discovering"}
-              >
-                Discover accounts
-              </EmptyButton>
-            </>
-            )
-        }
       </Empty>
       {hasHiddenAccounts && <HiddenAccountsBar />}
     </>
